+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 
+- Force closed alerts display after alert update.
 - Clean display preferences and profile rights table on uninstall
 
 ## [1.14.0] - 2026-04-30